integrate feature: whitelist
