+++ .gitmodules
@@ -19,3 +19,5 @@
 [submodule "lib/OZ-Upgradeable-Foundry"]
 	path = lib/OZ-Upgradeable-Foundry
 	url = https://github.com/jordaniza/OZ-Upgradeable-Foundry
+	branch = v1.4.0
+
