--- docs/Oracle.md
@@ -1,40 +1,0 @@
-# Contract Oracle
-**Title:** 
-**Author:** 
-
-**Description:** 
-
----
-### Globals Variables and Mappings
-
----
-### Functions
-
-#### Methods `GetAccount(address _wallet) → uint256`  Visibility:  external
-
-**Description**: 
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| address | _wallet |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | Variable |
-
-
-
-
----
-### Modifiers
-
----
-### Events
-
----
-### Structs
-
----
-### Enums
