+++ docs/USDC.md
@@ -1,3 +1,5 @@
+# Solidity API
+
 # Contract USDC
 **Title:** 
 **Author:** 
@@ -7,53 +9,8 @@
 ---
 ### Globals Variables and Mappings
 
-#### Variable
-| Type | Name | Visibility | Description |
-| ---- | ---- | ---------- | ----------- |
-| address | admin |  |  |
-
 ---
 ### Functions
-
-#### Methods `mint(address to, uint256 amount)`  Visibility:  external
-
-**Description**: 
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| address | to |
-| uint256 | amount |
-
-
-
-
-
-#### Methods `decimals() → uint8`  Visibility:  public
-
-**Description**: 
-
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint8 | Variable |
-
-
-
-
-#### Methods `burn(uint256 amount)`  Visibility:  external
-
-**Description**: 
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | amount |
-
-
-
-
 
 ---
 ### Modifiers
@@ -66,3 +23,4 @@
 
 ---
 ### Enums
+
