+++ docs/lib/ERC4626.md
@@ -1,341 +1,16 @@
+# Solidity API
+
 # Contract ERC4626
-**Title:** 
-**Author:** 
+**Title:** ERC4626 Method Implementation
+**Author:** Alfredo Lopez / Calculum
 
 **Description:** 
 
 ---
 ### Globals Variables and Mappings
 
-#### Variable
-| Type | Name | Visibility | Description |
-| ---- | ---- | ---------- | ----------- |
-| contract IERC20MetadataUpgradeable | _asset |  |  |
-
 ---
 ### Functions
-
-#### Methods `__ERC4626_init(contract IERC20MetadataUpgradeable __asset, string _name, string _symbol, uint8 decimals_)`  Visibility:  internal
-
-**Description**: 
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| contract IERC20MetadataUpgradeable | __asset |
-| string | _name |
-| string | _symbol |
-| uint8 | decimals_ |
-
-
-
-
-
-#### Methods `__ERC4626_init_unchained(contract IERC20MetadataUpgradeable asset_, string _name, string _symbol, uint8 decimals_)`  Visibility:  internal
-
-**Description**: 
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| contract IERC20MetadataUpgradeable | asset_ |
-| string | _name |
-| string | _symbol |
-| uint8 | decimals_ |
-
-
-
-
-
-#### Methods `decimals() → uint8`  Visibility:  public
-
-**Description**: 
-
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint8 | Variable |
-
-
-
-
-#### Methods `asset() → address`  Visibility:  public
-
-**Description**: See {IERC4262-asset}
-
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| address | Variable |
-
-
-
-
-#### Methods `totalAssets() → uint256`  Visibility:  public
-
-**Description**: Returns the total amount of the underlying asset that is “managed” by Vault.
-
-- SHOULD include any compounding that occurs from yield.
-- MUST be inclusive of any fees that are charged against assets in the Vault.
-- MUST NOT revert.
-
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | Variable |
-
-
-
-
-#### Methods `convertToShares(uint256 assets) → uint256 shares`  Visibility:  public
-
-**Description**: See {IERC4262-convertToShares}
-
-Will revert if asserts > 0, totalSupply > 0 and totalAssets = 0. That corresponds to a case where any asset
-would represent an infinite amout of shares.
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | assets |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | shares |
-
-
-
-
-#### Methods `convertToAssets(uint256 shares) → uint256 assets`  Visibility:  public
-
-**Description**: See {IERC4262-convertToAssets}
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | shares |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | assets |
-
-
-
-
-#### Methods `maxDeposit(address) → uint256`  Visibility:  public
-
-**Description**: See {IERC4262-maxDeposit}
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| address |  |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | Variable |
-
-
-
-
-#### Methods `maxMint(address) → uint256`  Visibility:  public
-
-**Description**: See {IERC4262-maxMint}
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| address |  |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | Variable |
-
-
-
-
-#### Methods `maxWithdraw(address owner) → uint256`  Visibility:  public
-
-**Description**: See {IERC4262-maxWithdraw}
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| address | owner |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | Variable |
-
-
-
-
-#### Methods `maxRedeem(address owner) → uint256`  Visibility:  public
-
-**Description**: See {IERC4262-maxRedeem}
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| address | owner |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | Variable |
-
-
-
-
-#### Methods `previewDeposit(uint256 assets) → uint256`  Visibility:  public
-
-**Description**: See {IERC4262-previewDeposit}
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | assets |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | Variable |
-
-
-
-
-#### Methods `previewMint(uint256 shares) → uint256`  Visibility:  public
-
-**Description**: See {IERC4262-previewMint}
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | shares |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | Variable |
-
-
-
-
-#### Methods `previewWithdraw(uint256 assets) → uint256`  Visibility:  public
-
-**Description**: See {IERC4262-previewWithdraw}
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | assets |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | Variable |
-
-
-
-
-#### Methods `previewRedeem(uint256 shares) → uint256`  Visibility:  public
-
-**Description**: See {IERC4262-previewRedeem}
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | shares |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | Variable |
-
-
-
-
-#### Methods `deposit(uint256 assets, address receiver) → uint256`  Visibility:  public
-
-**Description**: See {IERC4262-deposit}
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | assets |
-| address | receiver |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | Variable |
-
-
-
-
-#### Methods `mint(uint256 shares, address receiver) → uint256`  Visibility:  public
-
-**Description**: See {IERC4262-mint}
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | shares |
-| address | receiver |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | Variable |
-
-
-
-
-#### Methods `withdraw(uint256 assets, address receiver, address owner) → uint256`  Visibility:  public
-
-**Description**: See {IERC4262-withdraw}
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | assets |
-| address | receiver |
-| address | owner |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | Variable |
-
-
-
-
-#### Methods `redeem(uint256 shares, address receiver, address owner) → uint256`  Visibility:  public
-
-**Description**: See {IERC4262-redeem}
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | shares |
-| address | receiver |
-| address | owner |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | Variable |
-
-
-
 
 ---
 ### Modifiers
@@ -348,3 +23,4 @@ would represent an infinite amout of shares.
 
 ---
 ### Enums
+
