+++ docs/mock/MockUpOracle.md
@@ -1,3 +1,5 @@
+# Solidity API
+
 # Contract MockUpOracle
 **Title:** 
 **Author:** 
@@ -10,50 +12,6 @@
 ---
 ### Functions
 
-#### Methods `constructor(address _wallet, contract IERC20Upgradeable asset)`  Visibility:  public
-
-**Description**: 
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| address | _wallet |
-| contract IERC20Upgradeable | asset |
-
-
-
-
-
-#### Methods `GetAccount(address _wallet) → uint256`  Visibility:  public
-
-**Description**: 
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| address | _wallet |
-
-**Outputs:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | Variable |
-
-
-
-
-#### Methods `SetAssetValue(uint256 _newValue)`  Visibility:  public
-
-**Description**: 
-
-**Arguments:**
-| Type | Name |
-| ---- | ---- |
-| uint256 | _newValue |
-
-
-
-
-
 ---
 ### Modifiers
 
@@ -65,3 +23,4 @@
 
 ---
 ### Enums
+
