+++ foundry.toml
@@ -3,20 +3,20 @@ src = './src'
 out = './out'
 libs = ['node_modules', 'lib']
 test = './test/foundry'
-solc_version = '0.8.13'
 cache_path  = 'cache_forge'
 remappings = [
     '@openzeppelin/=node_modules/@openzeppelin/',
     'ds-test/=lib/forge-std/lib/ds-test/src/',
     'forge-std/=lib/forge-std/src/',
     'OZ-Upgradeable-Foundry/=lib/OZ-Upgradeable-Foundry/src/',
-    'openzeppelin-contracts-upgradeable/=lib/openzeppelin-contracts-upgradeable/',
-    'v2-periphery/=lib/v2-periphery/contracts/'
+    '@openzeppelin-contracts-upgradeable/=lib/openzeppelin-contracts-upgradeable/',
+    '@uniswap/v2-periphery/=lib/v2-periphery/contracts/'
 ]
 optimizer = true
-optimizer_runs = 200
+optimizer_runs = 999
 
 # Fixed Version
+solc_version = '0.8.17'
 ffi = true
 via-ir = true
