+++ package.json
@@ -4,10 +4,10 @@
   "description": "Draft Vault based on ERC4626",
   "main": "hardhat.config.ts",
   "scripts": {
-    "test": "npx hardhat test",
+    "test": "npx hardhat test ./test/index.test.ts --network hardhat",
     "test-local": "npx hardhat test --network local",
     "lint": "prettier --list-different 'contracts/**/*.sol'",
-    "build": "rm -rf docs/ && npx hardhat clean && npx hardhat compile && npx solidity-docgen --solc-module solc -i ./contracts -t ./templates",
+    "build": "rm -rf docs/ && npx hardhat clean && npx hardhat compile && npx hardhat docgen",
     "hardhat": "npx hardhat",
     "deploy": "npx hardhat run ./scripts/deploy.ts",
     "coverage": "npx hardhat coverage",
@@ -29,57 +29,60 @@
   "author": "Alfredo Lopez / Calculum",
   "license": "MIT",
   "devDependencies": {
-    "@dydxprotocol/v3-client": "^1.11.0",
+    "@dydxprotocol/v3-client": "^1.14.2",
+    "@nomicfoundation/hardhat-chai-matchers": "^1.0.3",
+    "@nomicfoundation/hardhat-foundry": "^1.0.0",
+    "@nomicfoundation/hardhat-network-helpers": "^1.0.6",
+    "@nomicfoundation/hardhat-toolbox": "^1.0.2",
+    "@nomicfoundation/solidity-analyzer": "^0.0.3",
     "@nomiclabs/hardhat-ethers": "^2.1.1",
     "@nomiclabs/hardhat-etherscan": "^3.1.0",
     "@nomiclabs/hardhat-truffle5": "^2.0.6",
-    "@nomiclabs/hardhat-waffle": "^2.0.3",
-    "@nomiclabs/hardhat-web3": "^2.0.0",
     "@openzeppelin/contracts": "^4.7.3",
     "@openzeppelin/contracts-upgradeable": "^4.7.3",
-    "@openzeppelin/hardhat-upgrades": "^1.19.0",
-    "@openzeppelin/upgrades-core": "^1.16.1",
-    "@typechain/ethers-v5": "^10.1.0",
-    "@types/chai": "^4.3.1",
-    "@types/mocha": "^9.1.1",
-    "@types/node": "^12.20.52",
-    "@types/ws": "^8.5.3",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
-    "chai": "^4.3.6",
+    "@openzeppelin/hardhat-upgrades": "^1.20.0",
+    "@openzeppelin/upgrades-core": "^1.18.0",
+    "@typechain/ethers-v5": "^10.2.0",
+    "@typechain/hardhat": "^6.1.5",
+    "@types/chai": "^4.3.4",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^18.15.3",
+    "@types/ws": "^8.5.4",
+    "@typescript-eslint/eslint-plugin": "^5.55.0",
+    "@typescript-eslint/parser": "^5.55.0",
+    "chai": "^4.3.7",
     "chai-as-promised": "^7.1.1",
     "chai-bn": "^0.3.1",
-    "dotenv": "^16.0.1",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.26.0",
+    "dotenv": "^16.0.3",
+    "eslint": "^8.36.0",
+    "eslint-config-prettier": "^8.7.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^3.4.1",
-    "eslint-plugin-promise": "^5.2.0",
-    "ethereum-waffle": "^3.4.4",
-    "ethers": "^5.7.0",
-    "hardhat": "^2.10.2",
-    "hardhat-contract-sizer": "^2.5.1",
-    "hardhat-gas-reporter": "^1.0.8",
-    "minimist": "^1.2.6",
-    "moment": "^2.29.3",
-    "prettier": "^2.6.2",
-    "prettier-plugin-solidity": "^1.0.0-beta.13",
-    "solc": "0.8.13",
-    "solhint": "^3.3.7",
-    "solidity-coverage": "^0.7.0",
-    "solidity-docgen": "^0.5.16",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-promise": "^6.1.1",
+    "ethereum-waffle": "^4.0.10",
+    "ethers": "5.7.2",
+    "hardhat": "2.12.8-viair.0",
+    "hardhat-contract-sizer": "^2.8.0",
+    "hardhat-gas-reporter": "^1.0.9",
+    "minimist": "^1.2.8",
+    "moment": "^2.29.4",
+    "node-fetch": "^3.3.1",
+    "prettier": "^2.8.4",
+    "prettier-plugin-solidity": "^1.1.3",
+    "solc": "0.8.17",
+    "solidity-coverage": "^0.7.21",
+    "solidity-docgen": "^0.6.0-beta.26",
     "ts-generator": "^0.1.1",
-    "ts-mocha": "^8.0.0",
-    "ts-morph": "^9.0.0",
-    "ts-node": "^8.10.2",
-    "tsconfig-paths": "^3.9.0",
-    "typescript": "^4.6.3",
-    "web3": "^1.7.5"
+    "ts-mocha": "^10.0.0",
+    "ts-morph": "^17.0.1",
+    "ts-node": "^10.9.1",
+    "tsconfig-paths": "^4.1.2",
+    "typescript": "^4.9.5",
+    "web3": "^1.8.2"
   },
   "dependencies": {
-    "@typechain/hardhat": "^6.1.2",
-    "node-fetch": "^2.6.7"
+    "yarn-upgrade-all": "^0.7.2"
   }
 }
