--- remappings.txt
@@ -1,16 +1,0 @@
-@ensdomains/=node_modules/@ensdomains/
-@openzeppelin/=node_modules/@openzeppelin/
-@oz-upgradeable/=lib/OZ-Upgradeable-Foundry/lib/openzeppelin-contracts-upgradeable/contracts/
-@oz/=lib/OZ-Upgradeable-Foundry/lib/openzeppelin-contracts/contracts/
-@prb/test/=lib/OZ-Upgradeable-Foundry/lib/prb-test/src/
-@std/=lib/OZ-Upgradeable-Foundry/lib/forge-std/src/
-OZ-Upgradeable-Foundry/=lib/OZ-Upgradeable-Foundry/src/
-ds-test/=lib/forge-std/lib/ds-test/src/
-eth-gas-reporter/=node_modules/eth-gas-reporter/
-forge-std/=lib/forge-std/src/
-hardhat/=node_modules/hardhat/
-openzeppelin-contracts-upgradeable/=lib/openzeppelin-contracts-upgradeable/
-openzeppelin-contracts/=lib/openzeppelin-contracts/
-prb-test/=lib/OZ-Upgradeable-Foundry/lib/prb-test/src/
-solmate/=lib/solmate/src/
-v2-periphery/=lib/v2-periphery/contracts/
