+++ src/lib/ERC4626.sol
@@ -1,12 +1,17 @@
 // SPDX-License-Identifier: MIT
-pragma solidity ^0.8.13;
+pragma solidity ^0.8.17;
 
-import "openzeppelin-contracts-upgradeable/contracts/token/ERC20/ERC20Upgradeable.sol";
-import "openzeppelin-contracts-upgradeable/contracts/token/ERC20/utils/SafeERC20Upgradeable.sol";
-import "openzeppelin-contracts-upgradeable/contracts/utils/ContextUpgradeable.sol";
-import "openzeppelin-contracts-upgradeable/contracts/proxy/utils/Initializable.sol";
+import "@openzeppelin-contracts-upgradeable/contracts/token/ERC20/ERC20Upgradeable.sol";
+import "@openzeppelin-contracts-upgradeable/contracts/token/ERC20/utils/SafeERC20Upgradeable.sol";
+import "@openzeppelin-contracts-upgradeable/contracts/utils/ContextUpgradeable.sol";
+import "@openzeppelin-contracts-upgradeable/contracts/proxy/utils/Initializable.sol";
 import "./IERC4626.sol";
 
+/**
+ * @title ERC4626 Method Implementation
+ * @dev Implementation of Methods based on standard ERC4626
+ * @custom:a Alfredo Lopez / Calculum
+ */
 abstract contract ERC4626 is Initializable, ContextUpgradeable, ERC20Upgradeable, IERC4626 {
     IERC20MetadataUpgradeable internal _asset;
     uint8 private _decimals;
