+++ test/foundry/WorkFlowTest.t.sol
@@ -1,38 +1,40 @@
 // SPDX-License-Identifier: MIT
-pragma solidity ^0.8.13;
+pragma solidity ^0.8.17;
 
 import "forge-std/Test.sol";
 import {Helpers} from "../../src/CalculumVault.sol";
+import {USDC} from "../../src/USDC.sol";
+import {MockUpOracle} from "../../src/mock/MockUpOracle.sol";
+import {UUPSProxy} from "OZ-Upgradeable-Foundry/UpgradeUUPS.sol";
 import {BasicTest} from "./BasicTest.t.sol";
 
 contract WorkFlowTest is BasicTest {
-
     event PendingDeposit(
         address indexed caller, address indexed receiver, uint256 assets, uint256 estimationOfShares
     );
 
     event Transfer(address indexed from, address indexed to, uint256 amount);
- 
+
     function testEpoch0() public {
         vm.startPrank(deployer);
         // Move to after the Maintenance Time Post Maintenance
-        vm.warp(block.timestamp + MAINT_TIME_AFTER * 6);
+        vm.warp(block.timestamp + MAINT_TIME_AFTER + MAINT_TIME_BEFORE);
         vault.setEpochDuration(EPOCH_DURATION, MAINT_TIME_AFTER, MAINT_TIME_BEFORE);
-        vault.setInitialValue([
-            MIN_DEPOSIT_PER_ADDR,
-            MAX_DEPOSIT_PER_ADDR,
-            TOKEN_MAX_TOTAL_SUPPLY
-        ]);
+        vault.setInitialValue([MIN_DEPOSIT_PER_ADDR, MAX_DEPOSIT_PER_ADDR, TOKEN_MAX_TOTAL_DEPOSIT]);
         vm.stopPrank();
 
         // Test investor0: alice
         address alice = investors[0];
         vm.startPrank(alice);
         uint256 depositAmount = MAX_DEPOSIT_PER_ADDR + 1;
-        vm.expectRevert(abi.encodeWithSelector(Helpers.DepositExceededMax.selector, alice, MAX_DEPOSIT_PER_ADDR));
+        vm.expectRevert(
+            abi.encodeWithSelector(Helpers.DepositExceededMax.selector, alice, MAX_DEPOSIT_PER_ADDR)
+        );
         vault.deposit(depositAmount, alice);
         depositAmount = MIN_DEPOSIT_PER_ADDR - 1;
-        vm.expectRevert(abi.encodeWithSelector(Helpers.DepositAmountTooLow.selector, alice, depositAmount));
+        vm.expectRevert(
+            abi.encodeWithSelector(Helpers.DepositAmountTooLow.selector, alice, depositAmount)
+        );
         vault.deposit(depositAmount, alice);
         vm.stopPrank();
 
@@ -41,8 +43,11 @@ contract WorkFlowTest is BasicTest {
         oracle.setAssetValue(_usdc(1500));
         vm.startPrank(alice);
         uint256 balanceBefore = usdc.balanceOf(alice);
-        (Helpers.Status statusBefore, , , ) = vault.DEPOSITS(alice);
-        assertTrue(statusBefore == Helpers.Status.Inactive, "epoch 0: deposit status should be 0 before deposit");
+        (Helpers.Status statusBefore,,,) = vault.DEPOSITS(alice);
+        assertTrue(
+            statusBefore == Helpers.Status.Inactive,
+            "epoch 0: deposit status should be 0 before deposit"
+        );
         depositAmount = _usdc(1500);
         vm.expectEmit(true, true, false, true, address(usdc));
         emit Transfer(alice, address(vault), depositAmount);
@@ -50,14 +55,21 @@ contract WorkFlowTest is BasicTest {
         emit PendingDeposit(alice, alice, depositAmount, 1500 ether);
         vault.deposit(depositAmount, alice);
 
-        (Helpers.Status status, uint256 amountAssets , uint256 amountShares, uint256 finalAmount) = vault.DEPOSITS(alice);
-        assertTrue(status == Helpers.Status.Pending, "epoch 0: deposit status should be 1 when pending");
+        (Helpers.Status status, uint256 amountAssets, uint256 amountShares, uint256 finalAmount) =
+            vault.DEPOSITS(alice);
+        assertTrue(
+            status == Helpers.Status.Pending, "epoch 0: deposit status should be 1 when pending"
+        );
         assertEq(amountAssets, _usdc(1500), "epoch 0: wrong assets amount in vault");
         assertEq(amountShares, 1500 ether, "epoch 0: wrong shares amount in vault");
         assertEq(finalAmount, 0, "epoch 0: wrong final amount in vault");
         assertEq(vault.balanceOf(alice), 0, "epoch 0: wrong token balance when pending");
         assertEq(vault.balanceOf(alice), 0, "epoch 0: wrong token balance");
-        assertEq(usdc.balanceOf(alice) + depositAmount, balanceBefore, "epoch 0: wrong balance of usdc after deposit");
+        assertEq(
+            usdc.balanceOf(alice) + depositAmount,
+            balanceBefore,
+            "epoch 0: wrong balance of usdc after deposit"
+        );
 
         uint256 nextEpochTime = vault.getNextEpoch();
         vm.warp(nextEpochTime - 70 minutes);
@@ -142,7 +154,7 @@ contract WorkFlowTest is BasicTest {
         vault.dexTransfer();
 
         vm.expectEmit(true, true, false, true, address(usdc));
-        emit Transfer(address(vault), treasuryWallet, transferAmount);
+        emit Transfer(address(vault), treasuryWallet, 0);
         vault.feesTransfer();
         vm.stopPrank();
 
@@ -195,7 +207,7 @@ contract WorkFlowTest is BasicTest {
         vault.dexTransfer();
 
         vm.expectEmit(true, true, false, true, address(usdc));
-        emit Transfer(address(vault), treasuryWallet, 259500);
+        emit Transfer(address(vault), treasuryWallet, 0);
         vault.feesTransfer();
         vm.stopPrank();
